Nuevo

+++ database/migrations/2019_07_01_153802_create_reviews_table.php
@@ -14,10 +14,10 @@ class CreateReviewsTable extends Migration
     public function up()
     {
         Schema::create('reviews', function (Blueprint $table) {
-            $table->increments('id');
-            $table->unsignedInteger('clase_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('clase_id');
             $table->foreign('clase_id')->references('id')->on('clases');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')->references('id')->on('users'); 
             $table->integer('rating'); 
             $table->text('comentario')->nullable();
